+++ src/redturtle/volto/restapi/deserializer/blocks.py
@@ -9,8 +9,14 @@ from zope.component import adapter
 from zope.interface import implementer
 
 
-EXCLUDE_KEYS = ["@type", "token", "value", "@id", "query"]
-EXCLUDE_TYPES = ["title", "listing", "calendar", "searchEvents", "form"]
+EXCLUDE_KEYS = ["@type", "token", "value", "@id", "query", "bg_color"]
+EXCLUDE_TYPES = [
+    "title",
+    "listing",
+    "calendar",
+    "searchEvents",
+    "form",
+]
 
 
 class GenericResolveUIDDeserializer(object):